+++ bundle/src/main/java/com/adobe/acs/commons/fam/ThrottledTaskRunner.java
@@ -15,11 +15,13 @@
  */
 package com.adobe.acs.commons.fam;
 
+import aQute.bnd.annotation.ProviderType;
 import com.adobe.acs.commons.fam.mbean.ThrottledTaskRunnerMBean;
 /**
  * In addition to MBean operations, a ThrottledTaskRunner lets the caller schedule work and provides a throttle method.
  * The logCompletion method should also allow a runnable action provide appropriate notification of success/failure
  */
+@ProviderType
 public interface ThrottledTaskRunner extends ThrottledTaskRunnerMBean {
 
     /**
@@ -37,6 +39,14 @@ public interface ThrottledTaskRunner extends ThrottledTaskRunnerMBean {
      */
     public void scheduleWork(Runnable work);
 
+    /**
+     * Schedule some kind of work to run in the future using the internal thread pool.
+     * The work will be throttled according to the CPU/Memory settings.  This action can be canceled at any time.
+     * @param work 
+     * @param cancelHandler
+     */
+    public void scheduleWork(Runnable work, CancelHandler cancelHandler);
+    
     /**
      * Record statistics
      * @param created Task creation time (Milliseconds since epoch) -- This is when the work is added to the queue