+++ bundle/src/main/java/com/adobe/acs/commons/rewriter/impl/ResourceResolverMapTransformerFactory.java
@@ -23,12 +23,7 @@ import com.adobe.acs.commons.rewriter.AbstractTransformer;
 import com.adobe.acs.commons.util.ParameterUtil;
 import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.lang.StringUtils;
-import org.apache.felix.scr.annotations.Activate;
-import org.apache.felix.scr.annotations.Component;
-import org.apache.felix.scr.annotations.ConfigurationPolicy;
-import org.apache.felix.scr.annotations.Properties;
-import org.apache.felix.scr.annotations.Property;
-import org.apache.felix.scr.annotations.Service;
+import org.apache.felix.scr.annotations.*;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.commons.osgi.PropertiesUtil;
 import org.apache.sling.rewriter.ProcessingComponentConfiguration;
@@ -69,6 +64,7 @@ public final class ResourceResolverMapTransformerFactory implements TransformerF
     private Map<String, String[]> attributes;
     @Property(label = "Rewrite Attributes",
             description = "List of element/attribute pairs to rewrite",
+            cardinality = Integer.MAX_VALUE,
             value = {"img:src"})
     private static final String PROP_ATTRIBUTES = "attributes";
 
@@ -103,8 +99,17 @@ public final class ResourceResolverMapTransformerFactory implements TransformerF
 
     @Activate
     protected void activate(final Map<String, Object> config) {
-        final String[] attrProp = PropertiesUtil.toStringArray(config.get(PROP_ATTRIBUTES), DEFAULT_ATTRIBUTES);
-        this.attributes = ParameterUtil.toMap(attrProp, ":", ",");
+        final String[] test = PropertiesUtil.toStringArray(config.get(PROP_ATTRIBUTES), new String[]{});
+
+        String[] normalizedValue = PropertiesUtil.toStringArray(config.get(PROP_ATTRIBUTES), DEFAULT_ATTRIBUTES);
+
+        if (test.length == 1) {
+            if (StringUtils.contains(test[0], ",")) {
+                normalizedValue = StringUtils.split(test[0], ",");
+            }
+        }
+
+        this.attributes = ParameterUtil.toMap(normalizedValue, ":", ",");
     }
 
     public final class ResourceResolverMapTransformer extends AbstractTransformer {