+++ bundle/src/main/java/com/adobe/acs/commons/users/impl/ProtectedSystemUsers.java
@@ -135,6 +135,8 @@ public final class ProtectedSystemUsers {
                 "acs-commons-dispatcher-flush-service",
                 "acs-commons-twitter-updater-service",
                 "acs-commons-review-task-asset-mover-service",
+                "acs-commons-automatic-package-replicator-service",
+                "acs-commons-manage-controlled-processes-service"
     };
 
     public static final boolean isProtected(String principalName) {