+++ bundle/src/main/java/com/adobe/acs/commons/workflow/synthetic/SyntheticWorkflowModel.java
@@ -20,8 +20,11 @@
 
 package com.adobe.acs.commons.workflow.synthetic;
 
+import aQute.bnd.annotation.ProviderType;
+
 import java.util.Map;
 
+@ProviderType
 public interface SyntheticWorkflowModel {
 
     /**