+++ bundle/src/test/java/com/adobe/acs/commons/workflow/process/impl/ReplicateWithOptionsWorkflowProcessTest.java
@@ -1,0 +1,8 @@
+package com.adobe.acs.commons.workflow.process.impl;
+
+/**
+ * Created by dgonzale on 5/16/17.
+ */
+public class ReplicateWithOptionsWorkflowProcessTest {
+
+}