+++ CHANGELOG.md
@@ -11,16 +11,37 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
 ### Added
 
+- #916: AEM Assets Brand Portal workflow process and Agent filter
+- #924: Content Modification Framework API
+- #958: Named Image Transform Servlet Sharpen transform 
+- #1005: Asset Folder Properties Support to allow custom fields/properties to be added to AEM Assets Folders in UI
+- #1039: Health Check Status E-mailer
+- #1041: QR Code to Publish in Page Editor
+- #1067: Vanity Path Web server re-writer mapping 
+- Managed Controlled Processes framework with 4 sample tools: Folder Re-locator, Asset Report (space usage), Deep Prune, Asset Ingestor (aka AntEater v2)
 - `com.adobe.acs.commons.fam.actions.ActionsBatch` for bundling Fast Action Manager actions so multiple changes can be retried if any of them fail and break the commit.
+- Fast Action Manager now has a halt feature in the API which instantly stops an action manager and any of its scheduled work
 
 ### Changed
 
+- #1033: Allow Resource Resolver Map Factory's re-write attributes to be passed in as an array
 - Updated Fast Action Manager retry logic to support more failure cases properly.
-- Updated Fast Action Manager retry logic to  be savvy about interrupted exceptions thrown by the watchdog trying to kill the thread.
-
-<!---
+- Updated Fast Action Manager retry logic to be savvy about interrupted exceptions thrown by the watchdog trying to kill the thread.
 
 ### Fixed
+
+- #982: Fixed issue with Touch UI Icon Picker was prefixing icon classes with 'fa' 
+- #1008: Email subject mangled for non-latin chars
+- #1043: JCR Package Replication now populates the replicated by properties of the packaged resources with the actual user that requested the replication of the package (with configurable override via OSGi config for backwards compat) 
+- #1044: JCR Package Replication fixes a resource leak where the JCR Packages were not closed after being opened 
+- #1051: Emails sent via EmailService do not have connection/socket timeouts 
+- #1059: ResourceServiceManager no longer users admin resource resolver
+- #1064: Fixed NPE in ResourceServiceManager when no serviceReferences exist
+- Error page handler OSGi configuration missing web hint for 'not-found' behavior.
+- Touch UI Multi-field saved User-picker values were not populated in dialog 
+- Fast Action Manager is much more efficient in how it gauges CPU usage, which makes it even faster than before.
+
+<!---
  
 ### Removed